travel-booking-system/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
pydantic
python-jose
passlib
bcrypt==4.0.1
python-dotenv
httpx
openai
cryptography
email-validator
